packet: move action functions to packet files
