+++ src/detect.c
@@ -28,6 +28,7 @@
 #include "conf.h"
 
 #include "decode.h"
+#include "packet.h"
 #include "flow.h"
 #include "stream-tcp.h"
 #include "app-layer.h"
